requirements.txt
APScheduler==3.11.0
attrs==25.1.0
Automat==24.8.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
et_xmlfile==2.0.0
filelock==3.17.0
Flask==3.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
lxml==5.3.0
mailersend==0.5.8
MarkupSafe==3.0.2
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parsel==1.10.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
scrapy-user-agents==0.1.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
tldextract==5.1.3
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ua-parser==1.0.0
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
w3lib==2.2.1
Werkzeug==3.1.3
zope.interface==7.2
